Fixed Logger, Cleaned up Brainiac, added Load menu
Logger now correctly creates or appends to a log file. It correctly creates strings and the logged events are buffered to reduce disk output. Added a Load menu to the Start menu hookups with a Load and Back button. This stubout will allow me to update the load menu with any preexisting saves.